Update for Spring Boot 2.0

+++ spring-data-mongodb-datatables/src/main/java/org/springframework/data/mongodb/datatables/repository/DataTablesUtils.java
@@ -256,7 +256,7 @@ public class DataTablesUtils {
         }
 
         @Override
-        public int getOffset() {
+        public long getOffset() {
             return offset;
         }
